+++ tests/__init__.py
@@ -1,1 +1,0 @@
-"""ConnectionPool 单元测试包"""