+++ tests/conftest.py
@@ -1,6 +1,0 @@
-import os, sys
-
-os.chdir(os.path.dirname(os.path.abspath(__file__)))
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-from aim_nats.conftest import *  # noqa